Create a script to compile a C file without linking
This script takes the name of a C file stored in the variable $CFILE and compiles
it without performing the linking step. The output object file is named the same
as the C file but with the '.o' extension. For example, if the C file is 'main.c',
the output file will be 'main.o'. This script is useful for generating object files
for further linking or building projects.