Create a script to generate assembly code from a C file
This script takes the name of a C file stored in the variable $CFILE, runs it
through the C compiler's preprocessor to generate assembly code, and saves
the assembly code in an output file. The output file is named the same as the
C file but with the '.s' extension. For example, if the C file is 'main.c',
the output file will be 'main.s'. This script is useful for inspecting the
generated assembly code for debugging or analysis purposes.