Create a script to compile a C file into an executable
This script takes the name of a C file stored in the variable $CFILE and compiles
it into an executable named "cisfun." It uses the C compiler to perform the
compilation. After running this script, you will have an executable program
named "cisfun" that can be executed from the command line.