Create a C program to display the sizes of various types
This C program calculates and prints the sizes of various types, including int,
long, long long, char, float, and double. It is designed to provide the same
output as the example, even allowing for warnings. The program returns 0 upon
successful execution. Note that on some Linux systems, you may need to install
the 'libc6-dev-i386' package to test the '-m32' gcc option for 32-bit
compilation.