Create a script to generate Intel-syntax assembly code from C source
This script takes a C source file, specified by the variable $CFILE, and
generates Intel-syntax assembly code from it. The resulting assembly code is
saved in an output file with the same name as the C file but with the .s
extension. For example, if the C file is 'main.c,' the assembly code will be
saved in 'main.s.'