Create a C program to determine the last digit of a random number
This C program assigns a random number to the variable 'n' each time it is executed.
It uses the rand, srand, and RAND_MAX functions to generate a random value.
The program then extracts the last digit of the number stored in the variable 'n'
and prints it along with a corresponding message. Depending on the last digit's
value, the program outputs whether it's greater than 5, equal to 0, or less than 6
but not 0, followed by a new line.