Write a program that adds positive numbers.
This commit implements a program that adds positive numbers passed as arguments.
It prints the result followed by a new line. If no numbers are passed, it prints 0 followed by a new line.
If any of the numbers contain non-digit symbols, it prints "Error" followed by a new line and returns an exit code of 1.