Write a program that calculates the minimum number of coins to make change.
This commit implements a program that, when executed with the syntax './change cents', calculates the minimum number of coins required to make change for the specified 'cents' amount. If the number of arguments passed to the program is not exactly 1, it prints "Error" followed by a new line and returns an exit code of 1. The program uses `atoi` to parse the parameter passed to it. If the number passed as the argument is negative, it prints "0" followed by a new line. It can use an unlimited number of coins with values of 25, 10, 5, 2, and 1 cent to make the change.